Relocate localization tests to correct test file
