+++ Tests/AcknowExampleTests/AcknowListViewControllerTests.swift
@@ -9,7 +9,7 @@
 import UIKit
 import XCTest
 
-import AcknowList
+@testable import AcknowList
 
 class AcknowListViewControllerTests: XCTestCase {
 
@@ -24,4 +24,47 @@ class AcknowListViewControllerTests: XCTestCase {
         let cell = viewController.tableView(viewController.tableView, cellForRowAtIndexPath: NSIndexPath(forRow: 0, inSection: 0))
         XCTAssertEqual(cell.textLabel?.text, "AcknowList")
     }
+    
+    func testLocalizedStringReturnsEnglish() {
+        
+        class MockAcknowListViewController: AcknowListViewController {
+            
+            override class func preferredLanguageCode() -> String {
+                return "en-GB"
+            }
+        }
+        
+        XCTAssertEqual("Acknowledgements", MockAcknowListViewController.localizedString(forKey: "VTAckAcknowledgements", defaultString: ""))
+        
+        XCTAssertEqual("Generated by CocoaPods", MockAcknowListViewController.localizedString(forKey: "VTAckGeneratedByCocoaPods", defaultString: ""))
+    }
+    
+    func testLocalizedStringReturnsFrench() {
+        
+        class MockAcknowListViewController: AcknowListViewController {
+            
+            override class func preferredLanguageCode() -> String {
+                return "fr-FR"
+            }
+        }
+        
+        XCTAssertEqual("Remerciements", MockAcknowListViewController.localizedString(forKey: "VTAckAcknowledgements", defaultString: ""))
+        
+        XCTAssertEqual("Généré par CocoaPods", MockAcknowListViewController.localizedString(forKey: "VTAckGeneratedByCocoaPods", defaultString: ""))
+    }
+    
+    func testLocalizedStringReturnsSimplifiedChinese() {
+        
+        class MockAcknowListViewController: AcknowListViewController {
+            
+            override class func preferredLanguageCode() -> String {
+                return "zh-Hans"
+            }
+        }
+        
+        XCTAssertEqual("致谢", MockAcknowListViewController.localizedString(forKey: "VTAckAcknowledgements", defaultString: ""))
+        
+        XCTAssertEqual("使用 CocoaPods 生成", MockAcknowListViewController.localizedString(forKey: "VTAckGeneratedByCocoaPods", defaultString: ""))
+    }
+
 }
