+++ Tests/AcknowExampleTests/AcknowViewControllerTests.swift
@@ -21,47 +21,4 @@ class AcknowViewControllerTests: XCTestCase {
         XCTAssertEqual(viewController.title, "Title")
         XCTAssertEqual(viewController.textView!.text, "Text...")
     }
-    
-    func testLocalizedStringReturnsEnglish() {
-        
-        class MockAcknowListViewController: AcknowListViewController {
-                        
-            override class func preferredLanguageCode() -> String {
-                return "en-GB"
-            }
-        }
-        
-        XCTAssertEqual("Acknowledgements", MockAcknowListViewController.localizedString(forKey: "VTAckAcknowledgements", defaultString: ""))
-        
-        XCTAssertEqual("Generated by CocoaPods", MockAcknowListViewController.localizedString(forKey: "VTAckGeneratedByCocoaPods", defaultString: ""))
-    }
-    
-    func testLocalizedStringReturnsFrench() {
-        
-        class MockAcknowListViewController: AcknowListViewController {
-            
-            override class func preferredLanguageCode() -> String {
-                return "fr-FR"
-            }
-        }
-        
-        XCTAssertEqual("Remerciements", MockAcknowListViewController.localizedString(forKey: "VTAckAcknowledgements", defaultString: ""))
-        
-        XCTAssertEqual("Généré par CocoaPods", MockAcknowListViewController.localizedString(forKey: "VTAckGeneratedByCocoaPods", defaultString: ""))
-    }
-    
-    func testLocalizedStringReturnsSimplifiedChinese() {
-        
-        class MockAcknowListViewController: AcknowListViewController {
-            
-            override class func preferredLanguageCode() -> String {
-                return "zh-Hans"
-            }
-        }
-        
-        XCTAssertEqual("致谢", MockAcknowListViewController.localizedString(forKey: "VTAckAcknowledgements", defaultString: ""))
-        
-        XCTAssertEqual("使用 CocoaPods 生成", MockAcknowListViewController.localizedString(forKey: "VTAckGeneratedByCocoaPods", defaultString: ""))
-    }
-
 }
